Updated intro text regarding 'no user interaction required'
Intro section regarding 'no user interaction' required is not accurate. Not sure if we need to add a link to the Apple MDM documentation which confirms this? https://developer.apple.com/business/documentation/MDM-Protocol-Reference.pdf (Page 76):

"On iOS 10.3 and later, supported Software Update commands require supervision but not DEP enrollment. If there is
a passcode on the device, a user must enter it to start a software update."

I've replaced/updated our text so that it is more accurate, and aligns with the Apple documentation